+++ app/Http/Controllers/Admin/FileController.php
@@ -2,41 +2,34 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Illuminate\Http\Request;
+
 use App\Models\File as FileModel;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\FileRequest;
 
-use URL;
 use File;
-use View;
-use Config;
-use Storage;
-use Request;
-use Response;
-use Redirect;
 
 class FileController extends Controller {
-	
-	const UPLOAD_PATH = 'app/uploads';
 	
 	public function index()
 	{
 		$files = FileModel::paginate(20);
 		
-		return View::make('admin.files.index', compact('files'));
+		return view('admin.files.index', compact('files'));
 	}
 	
 	public function get($date, $name)
 	{
 		$file_path = str_replace('-', '/', $date) . '/' . $name;
-		$file_path = storage_path(self::UPLOAD_PATH . '/' . $file_path);
+		$file_path = storage_path(FileModel::getUploadPath() . '/' . $file_path);
 
 		if (File::exists($file_path) == FALSE)
 			throw new \Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 		
 		$file_type = File::type($file_path);
 	
-		return Response::make(File::get($file_path), 200)->header('Content-Type', $file_type);
+		return response(File::get($file_path), 200)->header('Content-Type', $file_type);
 	}
 
 	public function store(FileRequest $request)
@@ -44,7 +37,7 @@ class FileController extends Controller {
 		$file = $request->file('file');
 		$time = time();
 		
-		$path = storage_path(self::UPLOAD_PATH . date('/Y/m/d', $time));
+		$path = storage_path(FileModel::getUploadPath() . date('/Y/m/d', $time));
 		
 		if (File::isDirectory($path) == FALSE)
 			File::makeDirectory($path, 0755, TRUE);
@@ -63,21 +56,35 @@ class FileController extends Controller {
 		
 		$data = [
 			'id'  => $file->id,
-			'url' => URL::route('file_get', [date('Y-m-d', $time), $file->name])
+			'url' => route('file_get', [date('Y-m-d', $time), $file->name]),
+			'description' => $file->description
 		];
 		
 		return $this->ajaxResponse(['error' => FALSE, 'data' => $data]);
 	}
 	
-	public function delete($file_id)
+	public function update($file_id, Request $request)
+	{
+		$file = FileModel::find($file_id);
+
+		if (is_null($request->description) == FALSE)
+			$file->update(['description' => $request->description]);
+		
+		if ($request->ajax())
+			return $this->ajaxResponse(['error' => FALSE]);
+		else
+			return redirect()->back();
+	}
+
+	public function delete($file_id, Request $request)
 	{
 		$file = FileModel::find($file_id);
 		
 		$file->delete();
 		
-		if (Request::ajax())
+		if ($request->ajax())
 			return $this->ajaxResponse(['error' => FALSE]);
 		else
-			return Redirect::back();
+			return redirect()->back();
 	}
 }