+++ app/Http/Middleware/PortfolioCheck.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Closure;
+use Session;
+class PortfolioCheck
+{
+	public function handle($request, Closure $next)
+	{
+		if ($time = $request->session()->get('portfolio')) {
+			if (86400 >= (time() - $time))
+				return $next($request);
+		}
+
+		$request->session()->forget('portfolio');
+
+		return redirect()->route('portfolio_show_authorize');
+	}
+}