+++ resources/views/admin/post/edit.blade.php
@@ -139,5 +139,5 @@
 </form>
 @stop
 @section('scripts')
-<script src="{{ asset('/assets/admin/app.js') }}"></script>
+<script src="{{ asset('/assets/admin/posts.js') }}"></script>
 @stop